+++ examples/vips-magick.php
@@ -1,9 +1,14 @@
 #!/usr/bin/env php
 <?php
 
-require __DIR__ . '/vendor/autoload.php';
+require dirname(__DIR__) . '/vendor/autoload.php';
 
 use Jcupitt\Vips;
+
+if (count($argv) != 3) {
+    echo("usage: ./vips-magick.php input-image output-image\n");
+    exit(1);
+}
 
 /* Load an image with libvips, render to a large memory buffer, wrap a imagick
  * image around that, then use imagick to save as another file.